Cast generalFlag to byte for clarity

+++ src/main/java/org/dragonet/proxy/network/translator/pc/PCMultiBlockChangePacketTranslator.java
@@ -11,7 +11,7 @@ public class PCMultiBlockChangePacketTranslator implements PCPacketTranslator<Se
    public PEPacket[] translate(UpstreamSession session, ServerMultiBlockChangePacket packet) {
       UpdateBlockPacket pk = new UpdateBlockPacket();
       pk.records = new UpdateBlockPacket.UpdateBlockRecord[packet.getRecords().length];
-      byte generalFlag = packet.getRecords().length > 64 ? 8 : 3;
+      byte generalFlag = packet.getRecords().length > 64 ? (byte) 8 : (byte) 3;
 
       for(int i = 0; i < pk.records.length; ++i) {
          pk.records[i] = new UpdateBlockPacket.UpdateBlockRecord();
